Final Update of project
Few days ago i was corretting on it but i was unable to solve the problems i've found, an so i didn't upload before.
Even after i consulted with Bongioanni , we haven't solved it , but anyway in someway i had to do some compromise.
Corrected everything in such way to work this project.
One thing that doesn't work as i wanted is the incremental search; if you see the code , it's written in the comment the reason.